% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCOILWTICO_level.R
\docType{data}
\name{DCOILWTICO_level}
\alias{DCOILWTICO_level}
\title{Quarterly Average WTI Crude Oil Price (Level)}
\format{
A data frame with 161 observations and 2 variables:
\describe{
  \item{date}{Date of observation (class \code{Date})}
  \item{DCOILWTICO_level}{Quarterly average oil price in U.S. dollars per barrel}
}
}
\source{
Federal Reserve Economic Data (FRED),
Federal Reserve Bank of St. Louis.
Data downloaded on 2026-01-27.
\url{https://fred.stlouisfed.org/}
}
\usage{
DCOILWTICO_level
}
\description{
Quarterly average spot price of West Texas Intermediate (WTI)
crude oil at Cushing, Oklahoma.
}
\keyword{datasets}
