% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_weights.R
\name{fct_weights}
\alias{fct_weights}
\title{Conditional Forecast Variable Weights}
\usage{
fct_weights(fit, cond_var, target_var, horizon, p0 = 10000)
}
\arguments{
\item{fit}{An object of class \code{varest} (from \pkg{vars}) or \code{bvar} (from \pkg{BVAR})}

\item{cond_var}{A vector indicating which columns of `y` are conditionally constrained}

\item{target_var}{Column index of the target variable being forecast}

\item{horizon}{Forecast horizon (number of future periods)}

\item{p0}{diagonal element of the initial state covariance, with default of 1e4}
}
\value{
A list of weight matrices for each horizon step
}
\description{
Computes observation weights used in the Kalman smoother for a given target variable,
showing the relative contribution of observed variables to the conditional forecast.
}
\examples{

library(cforecast)
library(vars)
data(fred_macro)
# Fit a  VAR model
fit <- VAR(fred_macro[,-1], p = 2, type = "const")
# compute observation weights for PCEPILFE given the conditioning on DCOILWTICO
fct_weights_fm <- fct_weights(fit = fit, cond_var = 5, target_var = 2, horizon = 1)
# note that only DCOILWTICO's weight is nonzero for h=1

}
