% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cforecast_composition.R
\name{cforecast_composition}
\alias{cforecast_composition}
\title{Compute Conditional Forecast Composition}
\usage{
cforecast_composition(x, target_var)
}
\arguments{
\item{x}{cforecast result}

\item{target_var}{target variable}
}
\value{
a data.frame with the decomposition
}
\description{
Compute Conditional Forecast Composition
}
